requirements.txt
Django==5.0.6
Pillow==10.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
